deployment/requirements.txt
scikit-learn==1.2.2
numpy==1.26.4
pandas
streamlit
Pillow
streamlit_option_menu
statsmodels